splitting graphql-config utilities into new graphql-config-utilities package
