+++ packages/graphql-config-utilities/.npmignore
@@ -1,0 +1,3 @@
+src
+test
+tsconfig.json
