+++ packages/graphql-config-utilities/src/index.ts
@@ -1,0 +1,1 @@
+export * from './config';
