+++ packages/graphql-tool-utilities/src/index.ts
@@ -1,9 +1,1 @@
 export * from './ast';
-
-export {
-  getGraphQLProjectForSchemaPath,
-  getGraphQLProjectIncludedFilePaths,
-  getGraphQLProjects,
-  getGraphQLSchemaPaths,
-} from './config';
-export * from './utilities';
