--- packages/graphql-tool-utilities/src/utilities.ts
@@ -1,61 +1,0 @@
-import {existsSync} from 'fs';
-import {GraphQLProjectConfig} from 'graphql-config/lib/GraphQLProjectConfig';
-
-export const defaultGraphQLProjectName = 'GraphQL';
-
-// temporary utility until `graphql-config` supports this function natively
-// see: https://github.com/prisma/graphql-config/pull/113
-export function resolvePathRelativeToConfig(
-  project: GraphQLProjectConfig,
-  relativePath: string,
-) {
-  return project.resolveConfigPath(relativePath);
-}
-
-export function resolveProjectName(
-  project: GraphQLProjectConfig,
-  defaultName = defaultGraphQLProjectName,
-) {
-  return project.projectName || defaultName;
-}
-
-export function resolveSchemaPath(
-  project: GraphQLProjectConfig,
-  ignoreMissing = false,
-) {
-  // schemaPath is nullable in graphq-config even though it cannot actually be
-  // omitted. This function simplifies access to the schemaPath without
-  // requiring a type guard.
-  if (!project.schemaPath) {
-    // this case should never happen with a properly formatted config file.
-    // graphql-config currently does not perform any validation so it's possible
-    // for a mal-formed schema to be loaded at runtime.
-    throw new Error(
-      `Missing GraphQL schemaPath for project '${resolveProjectName(project)}'`,
-    );
-  }
-
-  // resolve fully qualified schemaPath
-  const schemaPath = project.resolveConfigPath(project.schemaPath);
-
-  if (ignoreMissing) {
-    return schemaPath;
-  }
-
-  if (!existsSync(schemaPath)) {
-    const forProject = project.projectName
-      ? ` for project '${project.projectName}'`
-      : '';
-    throw new Error(
-      [
-        `Schema not found${forProject}.`,
-        `Expected to find the schema at '${schemaPath}' but the path does not exist.`,
-        `Check '${
-          project.configPath
-        }' and verify that schemaPath is configured correctly${forProject}.`,
-      ].join(' '),
-    );
-  }
-
-  return schemaPath;
-}
